+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## Unreleased
+
+- Use a [configuration file](chainpulse.toml) instead of command-line arguments
+- Add support for listening on multiple chains simultaneously
+
 ## v0.1.2
 
 - Respond to SIGINT, SIGHUP and SIGTERM signals even when ran as PID 1, eg. in a distroless Docker container